pyproject.toml
[project]
name = "medical-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.47.0,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "llama-cpp-python (>=0.3.14,<0.4.0)",
    "torchxrayvision (>=1.3.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
